requirements.txt
# Core scraping
requests
beautifulsoup4
lxml
readability-lxml

# JS-rendered pages (for later strategies)
playwright

# Text handling
python-dateutil

# JSON & utilities
tqdm

# UI
streamlit

# Environment handling
python-dotenv
